requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.6
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.1.0
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgviz==1.7.5
install==1.3.5
Jinja2==3.1.3
kiwisolver==1.4.5
labelme==5.4.1
lazy_loader==0.3
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
multidict==6.0.5
natsort==8.4.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.1
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
QtPy==2.4.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
sympy==1.12
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
